+++ app/src/main/java/com/numero/material_gallery/components/list/ComponentListActivity.kt
@@ -19,6 +19,7 @@ import com.numero.material_gallery.components.appbar.top.TopAppBarTypeActivity
 import com.numero.material_gallery.components.DesignComponent
 import com.numero.material_gallery.components.button.MaterialButtonActivity
 import com.numero.material_gallery.components.button.MaterialButtonToggleGroupActivity
+import com.numero.material_gallery.components.card.MaterialCardActivity
 import com.numero.material_gallery.components.fab.ExtendedFabActivity
 import com.numero.material_gallery.components.fab.FabActivity
 import com.numero.material_gallery.components.selection.CheckboxActivity
